Optimized all strings to PROGMEM

+++ AssignmentExam/gameOneStone.cpp
@@ -1,6 +1,6 @@
 #include "gameOneStone.h"
 
-gameOneStone::gameOneStone(int _x, int _speed, int _size, int _maxFallLength){
+gameOneStone::gameOneStone(unsigned short _x, unsigned short _speed, unsigned short _size, unsigned short _maxFallLength){
     x = _x;
     y = 10;
     speed = _speed;
@@ -16,14 +16,14 @@ void gameOneStone::moveStep(){
     y += speed;
 }
 
-void gameOneStone::backToTop(int _newPos){
+void gameOneStone::backToTop(unsigned short _newPos){
     y = 10;
     if (++speed >= maxSpeed) speed = maxSpeed;
     if (speed == maxSpeed && ++size >= maxSize) size = maxSize;
     x = _newPos;
 }
 
-void gameOneStone::resetStone(int _newPos, int _speed, int _size){
+void gameOneStone::resetStone(unsigned short _newPos, unsigned short _speed, unsigned short _size){
     y = 10;
     speed = _speed;
     size = _size;
